src/test/java/com/tutorialspoint/app/InventoryControllerTest.java
package com.tutorialspoint.app;

import com.tutorialspoint.app.repository.InMemoryProductDao;
import com.tutorialspoint.app.service.SimpleProductManager;
import com.tutorialspoint.app.web.InventoryController;
import org.junit.Test;
import org.springframework.web.servlet.ModelAndView;

import java.util.ArrayList;
import java.util.Map;

import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertNotNull;

/**
 * Created by jose on 19/03/2017.
 */
public class InventoryControllerTest {
    @Test
    public void testHandleRequestView() throws Exception{
        InventoryController controller = new InventoryController();
        SimpleProductManager spm = new SimpleProductManager();
        spm.setProductDao(new InMemoryProductDao(new ArrayList<>()));
        controller.setProductManager(spm);
        // controller.setProductManager(new SimpleProductManager());
        ModelAndView modelAndView = controller.handleRequest(null, null);
        assertEquals("hello", modelAndView.getViewName());
        assertNotNull(modelAndView.getModel());
        @SuppressWarnings("unchecked")
        Map<String, Object> modelMap = (Map<String, Object>) modelAndView.getModel().get("model");
        String nowValue = (String) modelMap.get("now");
        assertNotNull(nowValue);
    }

}